requirements.txt
fastapi
uvicorn
pydantic
langchain-groq
langchain-core
langchain-community
chromadb
pandas
python-dotenv
requests
uuid
